add to table cours and equipement id user to user add in profile

+++ delete.php
@@ -1,11 +1,14 @@
 <?php
+
+$user_id = $_SESSION["id_user"];
+
 //patie delet cours f table
 if (isset($_GET['del_id'])) {
     $delet_cours = $_GET['del_id'];
-    $sqldelet = "DELETE FROM cours WHERE id_cours = $delet_cours";
+    $sqldelet = "DELETE FROM cours WHERE id_cours = $delet_cours AND id_user = $user_id";
 
     if (mysqli_query($conn, $sqldelet)) {
-        header("Location: index.php?succes=1");
+        header("Location: dashboard.php?succes=1");
         exit();
     } else {
         echo "Error:" . mysqli_error($conn);
@@ -14,10 +17,10 @@ if (isset($_GET['del_id'])) {
 //partie delet equipement in table
 if (isset($_GET['deleq_id'])) {
     $delet_equipement = $_GET['deleq_id'];
-    $sqldelet_equipement = "DELETE FROM equipement WHERE id_equipement= $delet_equipement";
+    $sqldelet_equipement = "DELETE FROM equipement WHERE id_equipement= $delet_equipement AND id_user = $user_id";
 
     if (mysqli_query($conn, $sqldelet_equipement)) {
-        header("Location: index.php?succes=1");
+        header("Location: dashboard.php?succes=1");
         exit();
     } else {
         echo "Error:" . mysqli_error($conn);
